+++ jgclark.Filer/src/index.js
@@ -3,7 +3,7 @@
 // -----------------------------------------------------------------------------
 // Plugin to help move selected pargraphs to other notes
 // Jonathan Clark
-// Last updated 17.8.2022, for v1.0.0-beta1
+// Last updated 18.3.2023, for v1.1.0-beta
 // -----------------------------------------------------------------------------
 
 // allow changes in plugin.json to trigger recompilation
@@ -21,7 +21,14 @@ export {
   moveParasToThisWeekly,
   moveParasToToday,
   moveParasToTomorrow,
-} from './fileItems'
+} from './moveItems'
+export {
+  copyNoteLinks,
+  copyRecentNoteLinks,
+  moveNoteLinks,
+  moveRecentNoteLinks,
+} from './noteLinks'
+export { updateSettings } from './filerHelpers'
 export { addIDAndAddToOtherNote } from './IDs'
 export { newNoteFromClipboard, newNoteFromSelection } from './newNote'
 
@@ -60,23 +67,6 @@ export async function onUpdateOrInstall(testUpdate: boolean = false): Promise<vo
 
     // Tell user the plugin has been updated
     await pluginUpdated(pluginJson, { code: updateSettingsResult, message: 'unused?' })
-
-    // Test to see if np.Globals has been installed
-    // TODO:
-    const allPlugins = DataStore.listPlugins()
-    const installedPlugins = DataStore.installedPlugins()
-    logDebug(pluginID, `- ${installedPlugins.length} / ${allPlugins.length}  plugins installed`)
-    const globalsInstalled = installedPlugins.filter((p) => p.id === 'np.Globals')
-    logDebug(pluginID, `- includes ${String(globalsInstalled.length)} np.Globals`)
-    if (globalsList.length === 0) {
-      // If not, ask user to install it
-      // TODO: check first
-      const globalPluginObjects = allPlugins.filter((p) => p.id === 'np.Globals')
-      const globalPluginObject = globalPluginObjects[0]
-      clo(globalPluginObject)
-      await DataStore.installPlugin(globalPluginObject, true)
-    }
-
     logDebug(pluginID, `- finished`)
 
   } catch (error) {